grönt test på not be possible to delete a post that belongs to someone else

+++ server/resources/posts/post-router.ts
@@ -53,14 +53,21 @@ postRouter.delete("/api/posts/:id", async (req, res) => {
       );
   }
   try {
-    const post = await PostModel.findByIdAndDelete(req.params.id);
+    const post = await PostModel.findById(req.params.id);
     if (!post) {
       return res.status(404).json(JSON.stringify({ message: `Post ${req.params.id} not found` }));
     }
+    
+    if (post.author.toString() !== req.session.user._id.toString()) {
+      return res.status(403).json(JSON.stringify({ message: "You are not authorized to update this post" }));
+    }
+  
+    await PostModel.findByIdAndDelete(req.params.id);
     res.status(204).json({ message: "Post deleted successfully" });
   } catch (err) {
     res.status(500).json(JSON.stringify({ message: "Could not delete post " }));
   }
-});
+}
+);
 
 export default postRouter;
